requirements.txt
asgiref==3.2.3
Django==3.0.3
djangorestframework==3.11.0
django-cors-headers==3.2.1
pytz==2019.3
sqlparse==0.3.0
django-import-export==2.2.0 